Add test for transposed Conv2D
