+++ TrainingExtensions/torch/src/python/aimet_torch/examples/test_models.py
@@ -42,6 +42,7 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
+from torch import nn as nn
 
 import aimet_torch.elementwise_ops as aimet_elementwise
 
@@ -790,3 +791,75 @@ class ModelWithDuplicateReLU(nn.Module):
         x = self.fc2(x).relu()
         x = self.fc3(x)
         return x
+
+
+class ModelWithTwoInputs(nn.Module):
+
+    def __init__(self):
+        super(ModelWithTwoInputs, self).__init__()
+        self.conv1_a = nn.Conv2d(1, 10, kernel_size=5)
+        self.maxpool1_a = nn.MaxPool2d(2)
+        self.relu1_a = nn.ReLU()
+
+        self.conv1_b = nn.Conv2d(1, 10, kernel_size=5)
+        self.maxpool1_b = nn.MaxPool2d(2)
+        self.relu1_b = nn.ReLU()
+
+        self.conv2 = nn.Conv2d(10, 20, kernel_size=5)
+        self.maxpool2 = nn.MaxPool2d(2)
+        self.relu2 = nn.LeakyReLU()
+        self.flatten = nn.Flatten()
+
+        self.fc1 = nn.Linear(320, 50)
+        self.relu3 = nn.ReLU()
+        self.dropout = nn.Dropout()
+        self.fc2 = nn.Linear(50, 10)
+
+        self.softmax = nn.LogSoftmax(dim=1)
+
+    def forward(self, x1, x2):
+        x1 = self.relu1_a(self.maxpool1_a(self.conv1_a(x1)))
+        x2 = self.relu1_b(self.maxpool1_b(self.conv1_b(x2)))
+        x = x1 + x2
+        x = self.relu2(self.maxpool2(self.conv2(x)))
+        x = self.flatten(x)
+        x = self.relu3(self.fc1(x))
+        x = self.dropout(x)
+        x = self.fc2(x)
+        return self.softmax(x)
+
+
+class ModelWithTransposeConv(nn.Module):
+
+    def __init__(self):
+        super(ModelWithTransposeConv, self).__init__()
+        self.conv1_a = nn.Conv2d(1, 10, kernel_size=5)
+        self.maxpool1_a = nn.MaxPool2d(2)
+        self.relu1_a = nn.ReLU()
+
+        self.conv1_b = nn.Conv2d(1, 10, kernel_size=5)
+        self.maxpool1_b = nn.MaxPool2d(2)
+        self.relu1_b = nn.ReLU()
+
+        self.conv2 = nn.ConvTranspose2d(10, 20, kernel_size=5)
+        self.maxpool2 = nn.MaxPool2d(2)
+        self.relu2 = nn.LeakyReLU()
+        self.flatten = nn.Flatten()
+
+        self.fc1 = nn.Linear(1280, 50)
+        self.relu3 = nn.ReLU()
+        self.dropout = nn.Dropout()
+        self.fc2 = nn.Linear(50, 10)
+
+        self.softmax = nn.LogSoftmax(dim=1)
+
+    def forward(self, x1, x2):
+        x1 = self.relu1_a(self.maxpool1_a(self.conv1_a(x1)))
+        x2 = self.relu1_b(self.maxpool1_b(self.conv1_b(x2)))
+        x = x1 + x2
+        x = self.relu2(self.maxpool2(self.conv2(x)))
+        x = self.flatten(x)
+        x = self.relu3(self.fc1(x))
+        x = self.dropout(x)
+        x = self.fc2(x)
+        return self.softmax(x)
