+++ TrainingExtensions/torch/test/python/test_qc_quantize_op.py
@@ -2,7 +2,7 @@
 # =============================================================================
 #  @@-COPYRIGHT-START-@@
 #
-#  Copyright (c) 2017-2018, Qualcomm Innovation Center, Inc. All rights reserved.
+#  Copyright (c) 2017-2022, Qualcomm Innovation Center, Inc. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are met:
@@ -187,127 +187,6 @@ class TestQcQuantizeOpStaticGrid:
         quant_out = quantizer.quantize(inp_tensor, MAP_ROUND_MODE_TO_PYMO['nearest'])
         expected_out = torch.tensor([0, 6, 75, 178, 181, 255], dtype=torch.float32)
         assert torch.equal(quant_out, expected_out)
-
-    def test_per_channel_symmetric_qdq(self):
-        """ Test tensor quantizer symmetric quantize-dequantize functionality on cpu """
-
-        quantizer = StaticGridPerChannelQuantizer(bitwidth=8, round_mode='nearest',
-                                                  quant_scheme=QuantScheme.post_training_tf,
-                                                  use_symmetric_encodings=True, enabled_by_default=True,
-                                                  num_channels=4)
-        encodings = [libpymo.TfEncoding() for _ in range(4)]
-        for index in range(3):
-            encodings[index].bw = 8
-            encodings[index].max = 3.81
-            encodings[index].min = -3.84
-            encodings[index].delta = 0.03
-            encodings[index].offset = -128
-
-        encodings[3].bw = 8
-        encodings[3].max = 6.35
-        encodings[3].min = -6.4
-        encodings[3].delta = 0.05
-        encodings[3].offset = -128
-
-        # delta is 0.040745098
-        quantizer.encoding = encodings
-
-        # Test quantize only on cpu
-        inp_tensor = torch.tensor([[-7, -5, -3, 0, .1, 2.5],
-                                   [-7, -5, -3, 0, .1, 2.5],
-                                   [-7, -5, -3, 0, .1, 2.5],
-                                   [-7, -5, -3, 0, .1, 2.5]],
-                                  dtype=torch.float32)
-
-        quant_out = quantizer.quantize_dequantize(inp_tensor, MAP_ROUND_MODE_TO_PYMO['nearest'])
-        expected_out = torch.tensor([[-3.84, -3.84, -3, 0, .089999996, 2.49],
-                                     [-3.84, -3.84, -3, 0, .089999996, 2.49],
-                                     [-3.84, -3.84, -3, 0, .089999996, 2.49],
-                                     [-6.4, -5, -3, 0, .1, 2.5]],
-                                    dtype=torch.float32)
-        assert torch.allclose(quant_out, expected_out, atol=1e-5)
-
-    def test_per_channel_asymmetric_qdq(self):
-        """ Test tensor quantizer asymmetric quantize-dequantize functionality on cpu """
-
-        quantizer = StaticGridPerChannelQuantizer(bitwidth=8, round_mode='nearest',
-                                                  quant_scheme=QuantScheme.post_training_tf,
-                                                  use_symmetric_encodings=False, enabled_by_default=True,
-                                                  num_channels=4)
-        encodings = [libpymo.TfEncoding() for _ in range(4)]
-        for index in range(3):
-            encodings[index].bw = 8
-            encodings[index].max = 1.9999956
-            encodings[index].min = -2.9999934
-            encodings[index].delta = 0.0196078
-            encodings[index].offset = -153
-
-        encodings[3].bw = 8
-        encodings[3].max = 2.404693
-        encodings[3].min = -5.995262
-        encodings[3].delta = 0.032941
-        encodings[3].offset = -182
-
-        # delta is 0.040745098
-        quantizer.encoding = encodings
-
-        # Test quantize only on cpu
-        inp_tensor = torch.tensor([[-7, -5, -3, 0, .1, 2.5],
-                                   [-7, -5, -3, 0, .1, 2.5],
-                                   [-7, -5, -3, 0, .1, 2.5],
-                                   [-7, -5, -3, 0, .1, 2.5]])
-
-        quant_out = quantizer.quantize_dequantize(inp_tensor, MAP_ROUND_MODE_TO_PYMO['nearest'])
-        expected_out = torch.tensor([[-3.0, -3.0, -3.0, 0, .098, 2.0],
-                                     [-3.0, -3.0, -3.0, 0, .098, 2.0],
-                                     [-3.0, -3.0, -3.0, 0, .098, 2.0],
-                                     [-5.9953, -5.0070, -2.9976, 0, .09888, 2.4047]],
-                                    dtype=torch.float32)
-        assert torch.allclose(quant_out, expected_out, atol=0.0001)
-
-    def test_per_channel_symmetric_compute_encodings(self):
-        """ Test tensor quantizer symmetric compute-encodings functionality on cpu """
-
-        quantizer = StaticGridPerChannelQuantizer(bitwidth=8, round_mode='nearest',
-                                                  quant_scheme=QuantScheme.post_training_tf,
-                                                  use_symmetric_encodings=True, enabled_by_default=True,
-                                                  num_channels=4)
-
-        inp_tensor = torch.tensor([[-7, -5, -3, 0, .1, 2.5],
-                                   [-5, -5, -3, 0, .1, 2.7],
-                                   [-6, -5, -3, 0, .1, 2.8],
-                                   [-5, -5, -3, 0, .1, 2]])
-        quantizer.update_encoding_stats(inp_tensor)
-        quantizer.compute_encoding()
-
-        assert len(quantizer.encoding) == 4
-        assert quantizer.encoding[0].max == 7
-        assert round(quantizer.encoding[0].min, 2) == -7.06
-
-        assert quantizer.encoding[3].max == 5
-        assert round(quantizer.encoding[3].min, 2) == -5.04
-
-    def test_per_channel_asymmetric_compute_encodings(self):
-        """ Test tensor quantizer asymmetric compute-encodings functionality on cpu """
-
-        quantizer = StaticGridPerChannelQuantizer(bitwidth=8, round_mode='nearest',
-                                                  quant_scheme=QuantScheme.post_training_tf,
-                                                  use_symmetric_encodings=False, enabled_by_default=True,
-                                                  num_channels=4)
-
-        inp_tensor = torch.tensor([[-7, -5, -3, 0, .1, 2.5],
-                                   [-5, -5, -3, 0, .1, 2.7],
-                                   [-6, -5, -3, 0, .1, 2.8],
-                                   [-5, -5, -3, 0, .1, 2]])
-        quantizer.update_encoding_stats(inp_tensor)
-        quantizer.compute_encoding()
-
-        assert len(quantizer.encoding) == 4
-        assert round(quantizer.encoding[0].max, 3) == 2.496
-        assert round(quantizer.encoding[0].min, 3) == -7.004
-
-        assert round(quantizer.encoding[3].max, 3) == 2.004
-        assert round(quantizer.encoding[3].min, 3) == -4.996
 
     def test_quantize_only_symmetric_signed_cpu(self):
         """ Test tensor quantizer quantize only symmetric signed functionality on cpu """
